+++ HospitalApp/Controllers/LabTestController.cs
@@ -3,6 +3,7 @@ using HospitalApp.Core.Application.Interfaces.Services;
 using HospitalApp.Core.Application.ViewModels.LabResult;
 using HospitalApp.Core.Application.ViewModels.LabTest;
 using HospitalApp.Core.Domain.Entities;
+using HospitalApp.Middlewares;
 using Microsoft.AspNetCore.Mvc;
 using static HospitalApp.Core.Application.ViewModels.LabTest.SelectLabTestsViewModel;
 
@@ -13,28 +14,45 @@ namespace HospitalApp.Controllers
         private readonly ILabTestService _service;
         private readonly ILabResultService _serviceResult;
         private readonly IAppointmentService _serviceAppointment;
+        private readonly ValidateUserSession _validateUserSession;
 
-        public LabTestController(ILabTestService service, ILabResultService serviceResult, IAppointmentService serviceAppointment)
+        public LabTestController(ILabTestService service, ILabResultService serviceResult, IAppointmentService serviceAppointment, ValidateUserSession validateUserSession)
         {
             _service = service;
             _serviceAppointment = serviceAppointment;
             _serviceResult = serviceResult;
+            _validateUserSession = validateUserSession;
         }
 
         public async Task<IActionResult> Index()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             var list = await _service.GetAllViewModel();
             return View(list);
         }
 
         public async Task<IActionResult> Create()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             return View("SaveLabTest", new SaveLabTestViewModel());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(SaveLabTestViewModel newLabTest)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveLabTest", newLabTest);
@@ -46,6 +64,11 @@ namespace HospitalApp.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             var labTestViewModel = await _service.GetByIdSaveViewModel(id);
 
             return View("SaveLabTest", labTestViewModel);
@@ -54,6 +77,11 @@ namespace HospitalApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SaveLabTestViewModel updatedlabTestViewModel)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveLabTest", updatedlabTestViewModel);
@@ -65,18 +93,33 @@ namespace HospitalApp.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             return View(await _service.GetByIdSaveViewModel(id));
         }
 
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             await _service.Delete(id);
             return RedirectToRoute(new { controller = "LabTest", action = "Index" });
         }
 
         public async Task<IActionResult> SelectLabTests(int AppointmentId)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             var availableLabTests = await _service.GetAvailableLabTestsAsync();
 
             var labTestSelections = availableLabTests.Select(labTest => new LabTestSelection
@@ -99,6 +142,11 @@ namespace HospitalApp.Controllers
         [HttpPost]
         public async Task<IActionResult> PerformLabTests(SelectLabTestsViewModel model)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             if (ModelState.IsValid)
             {
                 var labResultsToCreate = model.AvailableLabTests
